Reorganized solution.
