+++ Library/Board.cs
@@ -1,123 +1,10 @@
 ﻿using System;
-using static ThreeKings.MoveUtility;
-using static ThreeKings.Piece;
+using static ThreeKings.Utilities.MoveUtility;
+using static ThreeKings.Library.Piece;
 using static ThreeKings.MoveFinder;
 
-namespace ThreeKings
+namespace ThreeKings.Library
 {
-
-    public class Piece
-    {
-        public PieceType _type;
-        public bool hasMoved;
-
-        public enum PieceType
-        {
-            None = 0,
-            WhitePawn = 1,
-            WhiteKnight = 2,
-            WhiteBishop = 3,
-            WhiteRook = 4,
-            WhiteQueen = 5,
-            WhiteKing = 6,
-            BlackPawn = 7,
-            BlackKnight = 8,
-            BlackBishop = 9,
-            BlackRook = 10,
-            BlackQueen = 11,
-            BlackKing = 12
-        }
-
-        public Piece(PieceType type)
-        {
-            _type = type;
-            hasMoved = false;
-        }
-
-        public bool isWhite()
-        {
-            if (_type == PieceType.WhitePawn || _type == PieceType.WhiteKnight || _type == PieceType.WhiteBishop ||
-                _type == PieceType.WhiteRook || _type == PieceType.WhiteQueen || _type == PieceType.WhiteKing)
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        public bool isBlack()
-        {
-            if (_type == PieceType.BlackPawn || _type == PieceType.BlackKnight || _type == PieceType.BlackBishop ||
-                _type == PieceType.BlackRook || _type == PieceType.BlackQueen || _type == PieceType.BlackKing)
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        public bool isPiece()
-        {
-            if (_type == PieceType.None)
-            {
-                return false;
-            }
-            else
-            {
-                return true;
-            }
-        }
-    }
-
-    public enum Color
-    {
-        None = 0,
-        White = 1,
-        Black = 2
-    }
-
-    public class Move
-    {
-        public byte FromIndex;
-        public byte ToIndex;
-        public PieceType Promotion;
-
-        public Move(string uciMoveNotation)
-        {
-            string fromSquare = uciMoveNotation.Substring(0, 2);
-            string toSquare = uciMoveNotation.Substring(2, 2);
-            FromIndex = Notation.ToSquare(fromSquare);
-            ToIndex = Notation.ToSquare(toSquare);
-            Promotion = (uciMoveNotation.Length == 5) ? Notation.ToPiece(uciMoveNotation[4]) : PieceType.None;
-        }
-
-        public override string ToString()
-        {
-            //result represents the move in the long algebraic notation (without piece names)
-            string result = Notation.ToSquareName(FromIndex);
-            result += Notation.ToSquareName(ToIndex);
-            //the presence of a 5th character should mean promotion
-            if (Promotion != PieceType.None)
-                result += Notation.ToChar(Promotion);
-
-            return result;
-        }
-
-        public void Adjust()
-        {
-            if (isUpperEdge(ToIndex))
-            {
-                if (Promotion != PieceType.None)
-                {
-                    Promotion = Notation.ToPiece(Notation.ToChar(Promotion).ToString().ToUpper());
-                }
-            }
-        }
-    }
 
     public class Board
     {
@@ -847,4 +734,5 @@ namespace ThreeKings
         #endregion
 
     }
+
 }
