+++ Library/Enums/Color.cs
@@ -1,0 +1,9 @@
+﻿namespace ThreeKings.Library
+{
+    public enum Color
+    {
+        None = 0,
+        White = 1,
+        Black = 2
+    }
+}
