+++ Utilities/MoveUtility.cs
@@ -1,7 +1,8 @@
 ﻿using System;
-using static ThreeKings.Piece;
+using ThreeKings.Library;
+using static ThreeKings.Library.Piece;
 
-namespace ThreeKings
+namespace ThreeKings.Utilities
 {
     public class MoveUtility
     {
